refactoring astar - split classes - create Case.py - create Grid.py - add a config file - slit test and main script - update .gitignore
